requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.1
MarkupSafe==1.1.1
pupdb==0.1.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tabulate==0.8.7
urllib3==1.25.10
Werkzeug==1.0.1
win32-setctime==1.0.1
